chore(milestone-2-enhancements): Add proper seeding behavior
