+++ backend/question/drizzle/meta/_journal.json
@@ -6,7 +6,7 @@
       "idx": 0,
       "version": "7",
       "when": 1727595727110,
-      "tag": "0000_bizarre_midnight",
+      "tag": "0000_initial_schema",
       "breakpoints": true
     }
   ]
